Fix in progress application footer styles, fix seeding formatting, rename send confirmation email util file
