+++ lib/applications/sendConfirmationEmail.ts
