+++ pages/api/payment-received.ts
@@ -3,7 +3,7 @@ import { Prisma, ShopifyPaymentStatus } from '@prisma/client'; // Prisma client
 import prisma from '@prisma/index'; // Prisma client
 import crypto from 'crypto'; // Verifying Shopify Request
 import getRawBody from 'raw-body';
-import sendConfirmationEmail from '@lib/reports/utils';
+import sendConfirmationEmail from '@lib/applications/sendConfirmationEmail';
 import { stripPostalCode } from '@lib/utils/format';
 
 /**
