+++ prisma/dev-seed-utils/applicants.ts
@@ -84,7 +84,7 @@ const applicants: Array<UpsertApplicant> = [
         mspNumber: '12346',
         firstName: 'Doctor',
         lastName: 'Two',
-        phone: '234-567-8901',
+        phone: '2345678901',
         addressLine1: '163 University Ave.',
         addressLine2: null,
         city: 'Burnaby',
@@ -119,7 +119,7 @@ const applicants: Array<UpsertApplicant> = [
         mspNumber: '12346',
         firstName: 'Doctor',
         lastName: 'Two',
-        phone: '234-567-8901',
+        phone: '2345678901',
         addressLine1: '163 University Ave.',
         addressLine2: null,
         city: 'Burnaby',
@@ -163,7 +163,7 @@ const applicants: Array<UpsertApplicant> = [
         mspNumber: '12344',
         firstName: 'Doctor',
         lastName: 'Four',
-        phone: '234-567-8904',
+        phone: '2345678904',
         addressLine1: '164 University Ave.',
         addressLine2: null,
         city: 'Burnaby',
